GroupA-_CS462-main/requirements.txt
Flask>=3.0
joblib>=1.4
numpy>=2.1
Pillow>=11.0
scikit-image>=0.25
scikit-learn>=1.6
